+++ lib/determinator/tracking/rack/middleware.rb
@@ -16,7 +16,26 @@ module Determinator
           error = true
           raise
         ensure
-          Determinator::Tracking.finish!(status: status, error: !!error)
+          Determinator::Tracking.finish!(
+            status: status,
+            error: !!error,
+            endpoint: extract_endpoint(env)
+          )
+        end
+
+        private
+
+        def extract_endpoint(env)
+          parts = if params = env['action_dispatch.request.path_parameters']
+            [env['REQUEST_METHOD'], [params[:controller], params[:action]].join('#')]
+          elsif env['sinatra.route']
+            [env['sinatra.route']]
+          else
+            [env['REQUEST_METHOD']]
+          end
+          Determinator::Tracking.collect_endpoint_info(parts)
+        rescue
+          env['PATH_INFO']
         end
       end
     end