+++ spec/determinator/tracking/rack/middleware_spec.rb
@@ -6,7 +6,7 @@ describe Determinator::Tracking::Rack::Middleware do
   let(:headers) { {} }
   let(:response) { 'foo' }
   let(:app) { double(call: [status, headers, response]) }
-  let(:env) { double }
+  let(:env) { {'REQUEST_METHOD' => 'GET', 'PATH_INFO' => '/test'} }
   let(:subject) { described_class.new(app) }
 
   describe '#call' do
@@ -37,6 +37,50 @@ describe Determinator::Tracking::Rack::Middleware do
       it 'sets the status' do
         subject.call(env)
         expect(@test_request.attributes[:status]).to eq(status)
+      end
+
+      it 'sets the endpoint' do
+        subject.call(env)
+        expect(@test_request.endpoint).to eq('GET')
+      end
+
+      context 'with a rails request' do
+        let(:env) do
+          super().merge('action_dispatch.request.path_parameters' => {controller: 'foo', action: 'test'} )
+        end
+
+        it 'sets the endpoint using controller info' do
+          subject.call(env)
+          expect(@test_request.endpoint).to eq('GET foo#test')
+        end
+      end
+
+      context 'with a sinatra request' do
+        let(:env) do
+          super().merge('sinatra.route' => 'POST /foo/bar' )
+        end
+
+        it 'sets the endpoint using controller info' do
+          subject.call(env)
+          expect(@test_request.endpoint).to eq('POST /foo/bar')
+        end
+      end
+
+      context 'with endpoint_env_vars' do
+        before do
+          Determinator::Tracking.endpoint_env_vars = ['__DETERMINATOR_TEST_ENV_VAR']
+          ENV['__DETERMINATOR_TEST_ENV_VAR'] = 'foo'
+        end
+
+        after do
+          Determinator::Tracking.endpoint_env_vars = nil
+          ENV.delete('__DETERMINATOR_TEST_ENV_VAR')
+        end
+
+        it 'adds the info from env' do
+          subject.call(env)
+          expect(@test_request.endpoint).to eq('foo GET')
+        end
       end
 
       context 'when the request errors' do